requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1732732445543/work
annotated-types==0.7.0
anyio==4.8.0
archspec @ file:///croot/archspec_1709217642129/work
attrs==25.1.0
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.5.2
certifi @ file:///croot/certifi_1738623731865/work/certifi
cffi @ file:///croot/cffi_1726856441404/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
conda @ file:///croot/conda_1733841818148/work
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1733258728466/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1727775630457/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1731366181659/work
cryptography @ file:///croot/cryptography_1732129644106/work
distro @ file:///croot/distro_1714488253808/work
filetype==1.2.0
frozendict @ file:///croot/frozendict_1713194832637/work
frozenlist==1.5.0
google-ai-generativelanguage==0.6.16
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.69.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna @ file:///croot/idna_1714398848350/work
jiter==0.8.2
joblib==1.4.2
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
langchain==0.3.19
langchain-core==0.3.40
langchain-google-genai==2.0.11
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langgraph==0.3.2
langgraph-checkpoint==2.0.16
langgraph-prebuilt==0.1.1
langgraph-sdk==0.1.53
langsmith==0.3.11
libmambapy @ file:///croot/mamba-split_1732896306512/work/libmambapy
menuinst @ file:///croot/menuinst_1731364904452/work
msgpack==1.1.0
multidict==6.1.0
mypy @ file:///croot/mypy-split_1736791980643/work
mypy-extensions @ file:///work/perseverance-python-buildout/croot/mypy_extensions_1698863276135/work
nltk==3.9.1
numpy==2.2.3
openai==1.65.2
orjson==3.10.15
packaging @ file:///croot/packaging_1720101850331/work
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///croot/pluggy_1733169602837/work
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psutil @ file:///croot/psutil_1736367091698/work
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.10.6
pydantic_core==2.27.2
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
requests-toolbelt==1.0.0
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1727980156079/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1727769819935/work
setuptools==75.1.0
sniffio==1.3.1
SQLAlchemy==2.0.38
tenacity==9.0.0
textblob==0.19.0
tiktoken==0.9.0
tqdm @ file:///croot/tqdm_1724853939799/work
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
urllib3 @ file:///croot/urllib3_1727769808118/work
wheel==0.44.0
yarl==1.18.3
zstandard @ file:///croot/zstandard_1731356346222/work
